+++ 16-Modulo_Funcao-1/ex5.py
@@ -4,14 +4,3 @@
 #--- Leia um número do console e armazene em uma variável
 #--- Realize o calculo da raiz e armazene em uma segunda variável
 #--- Imprima o resultado e uma mensagem usando f-string (módulo 3)
-
-def calcularRaiz(radicando, indice):
-    return (radicando ** (1/indice))
-
-if __name__ == "__main__":
-    radicando = float(input("Insira o radicando: "))
-    indice = float(input("Insira o indice: "))
-
-    raiz = calcularRaiz(radicando, indice)
-
-    print(f'A raiz {indice}º de {radicando} é {raiz}')
